Bulky Changes to chat
Ports: https://github.com/tgstation/tgstation/pull/85257 and its fixes
