+++ html/statbrowser.html
@@ -1,3 +1,6 @@
-<ul id="menu" class="button-container"></ul>
-<div id="under_menu"></div>
-<div id="statcontent"></div>
+<div class="stat-container">
+  <div id="menu" class="button-container"></div>
+  <div id="under-menu"></div>
+  <div id="statcontent"></div>
+  <div id="under-content"></div>
+</div>
