+++ tgui/packages/tgui-panel/chat/ChatPageSettings.jsx
@@ -30,7 +30,23 @@ export const ChatPageSettings = (props) => {
   return (
     <Section>
       <Stack align="center">
-        <Stack.Item grow>
+        {!!!page.isMain && (
+          <Stack.Item>
+            <Button
+              color="blue"
+              icon="angles-left"
+              tooltip="Reorder tab to the left"
+              onClick={() =>
+                dispatch(
+                  moveChatPageLeft({
+                    pageId: page.id,
+                  }),
+                )
+              }
+            />
+          </Stack.Item>
+        )}
+        <Stack.Item grow ml={0.5}>
           <Input
             width="100%"
             value={page.name}
@@ -44,9 +60,24 @@ export const ChatPageSettings = (props) => {
             }
           />
         </Stack.Item>
+        {!!!page.isMain && (
+          <Stack.Item ml={0.5}>
+            <Button
+              color="blue"
+              icon="angles-right"
+              tooltip="Reorder tab to the right"
+              onClick={() =>
+                dispatch(
+                  moveChatPageRight({
+                    pageId: page.id,
+                  }),
+                )
+              }
+            />
+          </Stack.Item>
+        )}
         <Stack.Item>
           <Button.Checkbox
-            content="Mute"
             checked={page.hideUnreadCount}
             icon={page.hideUnreadCount ? 'bell-slash' : 'bell'}
             tooltip="Disables unread counter"
@@ -58,13 +89,15 @@ export const ChatPageSettings = (props) => {
                 }),
               )
             }
-          />
+          >
+            Mute
+          </Button.Checkbox>
         </Stack.Item>
-        {!page.isMain ? (
+        {!!!page.isMain && (
           <Stack.Item>
             <Button
-              icon="times"
               color="red"
+              icon="times"
               onClick={() =>
                 dispatch(
                   removeChatPage({
@@ -76,42 +109,6 @@ export const ChatPageSettings = (props) => {
               Remove
             </Button>
           </Stack.Item>
-        ) : (
-          ''
-        )}
-      </Stack>
-      <Divider />
-      <Stack align="center">
-        {!page.isMain ? (
-          <Stack.Item>
-            Reorder Chat:&emsp;
-            <Button
-              color="blue"
-              onClick={() =>
-                dispatch(
-                  moveChatPageLeft({
-                    pageId: page.id,
-                  }),
-                )
-              }
-            >
-              &laquo;
-            </Button>
-            <Button
-              color="blue"
-              onClick={() =>
-                dispatch(
-                  moveChatPageRight({
-                    pageId: page.id,
-                  }),
-                )
-              }
-            >
-              &raquo;
-            </Button>
-          </Stack.Item>
-        ) : (
-          ''
         )}
       </Stack>
       <Divider />
