+++ tgui/packages/tgui-panel/panelFocus.ts
@@ -7,9 +7,9 @@
  * @license MIT
  */
 
-import { vecLength, vecSubtract } from 'common/vector';
-import { canStealFocus, globalEvents } from 'tgui/events';
 import { focusMap } from 'tgui/focus';
+import { canStealFocus, globalEvents } from 'tgui/events';
+import { vecLength, vecSubtract } from 'common/vector';
 
 // Empyrically determined number for the smallest possible
 // text you can select with the mouse.
@@ -19,9 +19,9 @@ const deferredFocusMap = () => setTimeout(() => focusMap());
 
 export const setupPanelFocusHacks = () => {
   let focusStolen = false;
-  let clickStartPos = null;
+  let clickStartPos: number[] | null = null;
   window.addEventListener('focusin', (e) => {
-    focusStolen = canStealFocus(e.target);
+    focusStolen = canStealFocus(e.target as HTMLElement);
   });
   window.addEventListener('mousedown', (e) => {
     clickStartPos = [e.screenX, e.screenY];
@@ -31,6 +31,9 @@ export const setupPanelFocusHacks = () => {
       const clickEndPos = [e.screenX, e.screenY];
       const dist = vecLength(vecSubtract(clickEndPos, clickStartPos));
       if (dist >= MIN_SELECTION_DISTANCE) {
+        focusStolen = true;
+      }
+      if (document.activeElement?.className.includes('Button')) {
         focusStolen = true;
       }
     }
