--- tgui/packages/tgui-panel/settings/SettingsPanel.jsx
@@ -1,391 +1,0 @@
-/**
- * @file
- * @copyright 2020 Aleksej Komarov
- * @license MIT
- */
-
-import { toFixed } from 'common/math';
-import { capitalize } from 'common/string';
-import { useLocalState } from 'tgui/backend';
-import { useDispatch, useSelector } from 'tgui/backend';
-import {
-  Box,
-  Button,
-  Collapsible,
-  ColorBox,
-  Divider,
-  Input,
-  LabeledList,
-  NumberInput,
-  Section,
-  Stack,
-  Tabs,
-  TextArea,
-} from 'tgui/components';
-
-import { ChatPageSettings } from '../chat';
-import { clearChat, rebuildChat, saveChatToDisk } from '../chat/actions';
-import { THEMES } from '../themes';
-import {
-  addHighlightSetting,
-  changeSettingsTab,
-  removeHighlightSetting,
-  updateHighlightSetting,
-  updateSettings,
-} from './actions';
-import { FONTS, MAX_HIGHLIGHT_SETTINGS, SETTINGS_TABS } from './constants';
-import {
-  selectActiveTab,
-  selectHighlightSettingById,
-  selectHighlightSettings,
-  selectSettings,
-} from './selectors';
-
-export const SettingsPanel = (props) => {
-  const activeTab = useSelector(selectActiveTab);
-  const dispatch = useDispatch();
-  return (
-    <Stack fill>
-      <Stack.Item>
-        <Section fitted fill minHeight="8em">
-          <Tabs vertical>
-            {SETTINGS_TABS.map((tab) => (
-              <Tabs.Tab
-                key={tab.id}
-                selected={tab.id === activeTab}
-                onClick={() =>
-                  dispatch(
-                    changeSettingsTab({
-                      tabId: tab.id,
-                    }),
-                  )
-                }
-              >
-                {tab.name}
-              </Tabs.Tab>
-            ))}
-          </Tabs>
-        </Section>
-      </Stack.Item>
-      <Stack.Item grow={1} basis={0}>
-        {activeTab === 'general' && <SettingsGeneral />}
-        {activeTab === 'chatPage' && <ChatPageSettings />}
-        {activeTab === 'textHighlight' && <TextHighlightSettings />}
-      </Stack.Item>
-    </Stack>
-  );
-};
-
-export const SettingsGeneral = (props) => {
-  const { theme, fontFamily, fontSize, lineHeight } =
-    useSelector(selectSettings);
-  const dispatch = useDispatch();
-  const [freeFont, setFreeFont] = useLocalState('freeFont', false);
-  return (
-    <Section>
-      <LabeledList>
-        <LabeledList.Item label="Theme">
-          {THEMES.map((THEME) => (
-            <Button
-              key={THEME}
-              content={capitalize(THEME)}
-              selected={theme === THEME}
-              color="transparent"
-              onClick={() =>
-                dispatch(
-                  updateSettings({
-                    theme: THEME,
-                  }),
-                )
-              }
-            />
-          ))}
-        </LabeledList.Item>
-        <LabeledList.Item label="Font style">
-          <Stack.Item>
-            {(!freeFont && (
-              <Collapsible
-                title={fontFamily}
-                width={'100%'}
-                buttons={
-                  <Button
-                    icon={freeFont ? 'lock-open' : 'lock'}
-                    color={freeFont ? 'good' : 'bad'}
-                    onClick={() => {
-                      setFreeFont(!freeFont);
-                    }}
-                  >Custom font
-                  </Button>
-                }
-              >
-                {FONTS.map((FONT) => (
-                  <Button
-                    key={FONT}
-                    fontFamily={FONT}
-                    selected={fontFamily === FONT}
-                    color="transparent"
-                    onClick={() =>
-                      dispatch(
-                        updateSettings({
-                          fontFamily: FONT,
-                        }),
-                      )
-                    }
-                  >{FONT}
-                  </Button>
-                ))}
-              </Collapsible>
-            )) || (
-              <Stack>
-                <Input
-                  width={'100%'}
-                  value={fontFamily}
-                  onChange={(e, value) =>
-                    dispatch(
-                      updateSettings({
-                        fontFamily: value,
-                      }),
-                    )
-                  }
-                />
-                <Button
-                  ml={0.5}
-                  content="Custom font"
-                  icon={freeFont ? 'lock-open' : 'lock'}
-                  color={freeFont ? 'good' : 'bad'}
-                  onClick={() => {
-                    setFreeFont(!freeFont);
-                  }}
-                />
-              </Stack>
-            )}
-          </Stack.Item>
-        </LabeledList.Item>
-        <LabeledList.Item label="Font size" verticalAlign='middle'>
-          <Stack align="center">
-            <Stack.Item>
-              <Button icon="minus" onClick={() => dispatch(updateSettings({ fontSize: Math.max(fontSize - 1, 8) }))} />
-            </Stack.Item>
-            <Stack.Item>
-          <NumberInput
-            width="4.2em"
-            step={1}
-            stepPixelSize={5}
-            minValue={8}
-            maxValue={32}
-            value={fontSize}
-            unit="px"
-            format={(value) => toFixed(value)}
-            onChange={(value) =>
-              dispatch(
-                updateSettings({
-                  fontSize: value,
-                }),
-              )
-            }
-          />
-            </Stack.Item>
-            <Stack.Item>
-            <Button icon="plus" onClick={() => dispatch(updateSettings({ fontSize: Math.min(fontSize + 1, 32) }))} />
-            </Stack.Item>
-          </Stack>
-        </LabeledList.Item>
-        <LabeledList.Item label="Line height" verticalAlign='middle'>
-          <Stack align="center">
-            <Stack.Item>
-              <Button icon="minus" onClick={() => dispatch(updateSettings({ lineHeight: Math.max(lineHeight - 0.01, 0.8) }))} />
-            </Stack.Item>
-            <Stack.Item>
-          <NumberInput
-            width="4.2em"
-            step={0.01}
-            stepPixelSize={2}
-            minValue={0.8}
-            maxValue={5}
-            value={lineHeight}
-            format={(value) => toFixed(value, 2)}
-            onDrag={(value) =>
-              dispatch(
-                updateSettings({
-                  lineHeight: value,
-                }),
-              )
-            }
-          />
-            </Stack.Item>
-            <Stack.Item>
-            <Button icon="plus" onClick={() => dispatch(updateSettings({ lineHeight: Math.min(lineHeight + 0.01, 5) }))} />
-            </Stack.Item>
-          </Stack>
-        </LabeledList.Item>
-      </LabeledList>
-      <Divider />
-      <Stack fill>
-        <Stack.Item grow mt={0.15}>
-          <Button
-            content="Save chat log"
-            icon="save"
-            tooltip="Export current tab history into HTML file"
-            onClick={() => dispatch(saveChatToDisk())}
-          />
-        </Stack.Item>
-        <Stack.Item mt={0.15}>
-          <Button.Confirm
-            content="Clear chat"
-            icon="trash"
-            tooltip="Erase current tab history"
-            onClick={() => dispatch(clearChat())}
-          />
-        </Stack.Item>
-      </Stack>
-    </Section>
-  );
-};
-
-const TextHighlightSettings = (props) => {
-  const highlightSettings = useSelector(selectHighlightSettings);
-  const dispatch = useDispatch();
-  return (
-    <Section fill scrollable height="250px">
-      <Stack vertical>
-        {highlightSettings.map((id, i) => (
-          <TextHighlightSetting
-            key={i}
-            id={id}
-            mb={i + 1 === highlightSettings.length ? 0 : '10px'}
-          />
-        ))}
-        {highlightSettings.length < MAX_HIGHLIGHT_SETTINGS && (
-          <Stack.Item>
-            <Button
-              color="transparent"
-              icon="plus"
-              content="Add Highlight Setting"
-              onClick={() => {
-                dispatch(addHighlightSetting());
-              }}
-            />
-          </Stack.Item>
-        )}
-      </Stack>
-      <Divider />
-      <Box>
-        <Button icon="check" onClick={() => dispatch(rebuildChat())}>
-          Apply now
-        </Button>
-        <Box inline fontSize="0.9em" ml={1} color="label">
-          Can freeze the chat for a while.
-        </Box>
-      </Box>
-    </Section>
-  );
-};
-
-const TextHighlightSetting = (props) => {
-  const { id, ...rest } = props;
-  const highlightSettingById = useSelector(selectHighlightSettingById);
-  const dispatch = useDispatch();
-  const {
-    highlightColor,
-    highlightText,
-    highlightWholeMessage,
-    matchWord,
-    matchCase,
-  } = highlightSettingById[id];
-  return (
-    <Stack.Item {...rest}>
-      <Stack mb={1} color="label" align="baseline">
-        <Stack.Item grow>
-          <Button
-            content="Delete"
-            color="transparent"
-            icon="times"
-            onClick={() =>
-              dispatch(
-                removeHighlightSetting({
-                  id: id,
-                }),
-              )
-            }
-          />
-        </Stack.Item>
-        <Stack.Item>
-          <Button.Checkbox
-            checked={highlightWholeMessage}
-            content="Whole Message"
-            tooltip="If this option is selected, the entire message will be highlighted in yellow."
-            onClick={() =>
-              dispatch(
-                updateHighlightSetting({
-                  id: id,
-                  highlightWholeMessage: !highlightWholeMessage,
-                }),
-              )
-            }
-          />
-        </Stack.Item>
-        <Stack.Item>
-          <Button.Checkbox
-            content="Exact"
-            checked={matchWord}
-            tooltipPosition="bottom-start"
-            tooltip="If this option is selected, only exact matches (no extra letters before or after) will trigger. Not compatible with punctuation. Overriden if regex is used."
-            onClick={() =>
-              dispatch(
-                updateHighlightSetting({
-                  id: id,
-                  matchWord: !matchWord,
-                }),
-              )
-            }
-          />
-        </Stack.Item>
-        <Stack.Item>
-          <Button.Checkbox
-            content="Case"
-            tooltip="If this option is selected, the highlight will be case-sensitive."
-            checked={matchCase}
-            onClick={() =>
-              dispatch(
-                updateHighlightSetting({
-                  id: id,
-                  matchCase: !matchCase,
-                }),
-              )
-            }
-          />
-        </Stack.Item>
-        <Stack.Item>
-          <ColorBox mr={1} color={highlightColor} />
-          <Input
-            width="5em"
-            monospace
-            placeholder="#ffffff"
-            value={highlightColor}
-            onInput={(e, value) =>
-              dispatch(
-                updateHighlightSetting({
-                  id: id,
-                  highlightColor: value,
-                }),
-              )
-            }
-          />
-        </Stack.Item>
-      </Stack>
-      <TextArea
-        height="3em"
-        value={highlightText}
-        placeholder="Put words to highlight here. Separate terms with commas, i.e. (term1, term2, term3)"
-        onChange={(e, value) =>
-          dispatch(
-            updateHighlightSetting({
-              id: id,
-              highlightText: value,
-            }),
-          )
-        }
-      />
-    </Stack.Item>
-  );
-};
