+++ tgui/packages/tgui-panel/settings/constants.ts
@@ -18,6 +18,10 @@ export const SETTINGS_TABS = [
     id: 'chatPage',
     name: 'Chat Tabs',
   },
+  {
+    id: 'statPanel',
+    name: 'Stat Panel',
+  },
 ];
 
 export const FONTS_DISABLED = 'Default';
