+++ tgui/packages/tgui-panel/themes.ts
@@ -6,11 +6,20 @@
 
 export const THEMES = ['light', 'dark'];
 
-// warning: if you change this, also update /mob/verb/hotkey_toggle's detection
-// also /mob/living/silicon/robot/hotkey_toggle
-const COLOR_DARK_BG = '#202020';
-const COLOR_DARK_BG_DARKER = '#171717';
-const COLOR_DARK_TEXT = '#a4bad6';
+const COLORS = {
+  DARK: {
+    BG_BASE: '#202020',
+    BG_SECOND: '#171717',
+    BUTTON: '#494949',
+    TEXT: '#A4BAD6',
+  },
+  LIGHT: {
+    BG_BASE: '#EEEEEE',
+    BG_SECOND: '#FFFFFF',
+    BUTTON: 'none',
+    TEXT: '#000000',
+  },
+};
 
 let setClientThemeTimer: NodeJS.Timeout;
 
@@ -33,102 +42,53 @@ export const setClientTheme = (name) => {
     Byond.command(`.output statbrowser:set_theme ${name}`);
   }, 1500);
 
-  if (name === 'light') {
-    return Byond.winset({
-      // Main windows
-      'infowindow.background-color': 'none',
-      'infowindow.text-color': '#000000',
-      'info.background-color': '#FFFFFF',
-      'info.text-color': '#000000',
-      'browseroutput.background-color': 'none',
-      'browseroutput.text-color': '#000000',
-      'outputwindow.background-color': 'none',
-      'outputwindow.text-color': '#000000',
-      'mainwindow.background-color': 'none',
-      'split.background-color': 'none',
-      // Buttons
-      'changelog.background-color': 'none',
-      'changelog.text-color': '#000000',
-      'rules.background-color': 'none',
-      'rules.text-color': '#000000',
-      'wiki.background-color': 'none',
-      'wiki.text-color': '#000000',
-      'github.background-color': 'none',
-      'github.text-color': '#000000',
-      'discord.background-color': 'none',
-      'discord.text-color': '#000000',
-      'tickets.background-color': 'none',
-      'tickets.text-color': '#000000',
-      // Status and verb tabs
-      'output.background-color': 'none',
-      'output.text-color': '#000000',
-      'statwindow.background-color': 'none',
-      'statwindow.text-color': '#000000',
-      // Say, OOC, me Buttons etc.
-      'saybutton.background-color': 'none',
-      'saybutton.text-color': '#000000',
-      'oocbutton.background-color': 'none',
-      'oocbutton.text-color': '#000000',
-      'hotkey_toggle.background-color': 'none',
-      'hotkey_toggle.text-color': '#000000',
-      'mebutton.background-color': 'none',
-      'mebutton.text-color': '#000000',
-      'input.background-color': '#FFFFFF',
-      'input.text-color': '#000000',
-      // Asset cache/tooltip
-      'asset_cache_browser.background-color': 'none',
-      'asset_cache_browser.text-color': '#000000',
-      'tooltip.background-color': 'none',
-      'tooltip.text-color': '#000000',
-    });
+  const themeColor = COLORS[name.toUpperCase()];
+  if (!themeColor) {
+    return;
   }
-  if (name === 'dark') {
-    Byond.winset({
-      // Main windows
-      'infowindow.background-color': COLOR_DARK_BG,
-      'infowindow.text-color': COLOR_DARK_TEXT,
-      'info.background-color': COLOR_DARK_BG,
-      'info.text-color': COLOR_DARK_TEXT,
-      'browseroutput.background-color': COLOR_DARK_BG,
-      'browseroutput.text-color': COLOR_DARK_TEXT,
-      'outputwindow.background-color': COLOR_DARK_BG,
-      'outputwindow.text-color': COLOR_DARK_TEXT,
-      'mainwindow.background-color': COLOR_DARK_BG,
-      'split.background-color': COLOR_DARK_BG,
-      // Buttons
-      'changelog.background-color': '#494949',
-      'changelog.text-color': COLOR_DARK_TEXT,
-      'rules.background-color': '#494949',
-      'rules.text-color': COLOR_DARK_TEXT,
-      'wiki.background-color': '#494949',
-      'wiki.text-color': COLOR_DARK_TEXT,
-      'github.background-color': '#3a3a3a',
-      'github.text-color': COLOR_DARK_TEXT,
-      'discord.background-color': '#494949',
-      'discord.text-color': COLOR_DARK_TEXT,
-      'tickets.background-color': '#494949',
-      'tickets.text-color': COLOR_DARK_TEXT,
-      // Status and verb tabs
-      'output.background-color': COLOR_DARK_BG_DARKER,
-      'output.text-color': COLOR_DARK_TEXT,
-      'statwindow.background-color': COLOR_DARK_BG_DARKER,
-      'statwindow.text-color': COLOR_DARK_TEXT,
-      // Say, OOC, me Buttons etc.
-      'saybutton.background-color': COLOR_DARK_BG,
-      'saybutton.text-color': COLOR_DARK_TEXT,
-      'oocbutton.background-color': COLOR_DARK_BG,
-      'oocbutton.text-color': COLOR_DARK_TEXT,
-      'hotkey_toggle.background-color': COLOR_DARK_BG,
-      'hotkey_toggle.text-color': COLOR_DARK_TEXT,
-      'mebutton.background-color': COLOR_DARK_BG,
-      'mebutton.text-color': COLOR_DARK_TEXT,
-      'input.background-color': COLOR_DARK_BG_DARKER,
-      'input.text-color': COLOR_DARK_TEXT,
-      // Asset cache/tooltip
-      'asset_cache_browser.background-color': COLOR_DARK_BG,
-      'asset_cache_browser.text-color': COLOR_DARK_TEXT,
-      'tooltip.background-color': COLOR_DARK_BG,
-      'tooltip.text-color': COLOR_DARK_TEXT,
-    });
-  }
+
+  return Byond.winset({
+    // Main windows
+    'infowindow.background-color': themeColor.BG_BASE,
+    'infowindow.text-color': themeColor.TEXT,
+    'info.background-color': themeColor.BG_BASE,
+    'info.text-color': themeColor.TEXT,
+    'browseroutput.background-color': themeColor.BG_BASE,
+    'browseroutput.text-color': themeColor.TEXT,
+    'outputwindow.background-color': themeColor.BG_BASE,
+    'outputwindow.text-color': themeColor.TEXT,
+    'mainwindow.background-color': themeColor.BG_BASE,
+    'split.background-color': themeColor.BG_BASE,
+    // Buttons
+    'changelog.background-color': themeColor.BUTTON,
+    'changelog.text-color': themeColor.TEXT,
+    'rules.background-color': themeColor.BUTTON,
+    'rules.text-color': themeColor.TEXT,
+    'wiki.background-color': themeColor.BUTTON,
+    'wiki.text-color': themeColor.TEXT,
+    'forum.background-color': themeColor.BUTTON,
+    'forum.text-color': themeColor.TEXT,
+    'github.background-color': themeColor.BUTTON,
+    'github.text-color': themeColor.TEXT,
+    'report-issue.background-color': themeColor.BUTTON,
+    'report-issue.text-color': themeColor.TEXT,
+    'fullscreen-toggle.background-color': themeColor.BUTTON,
+    'fullscreen-toggle.text-color': themeColor.TEXT,
+    // Status and verb tabs
+    'output.background-color': themeColor.BG_BASE,
+    'output.text-color': themeColor.TEXT,
+    // Say, OOC, me Buttons etc.
+    'saybutton.background-color': themeColor.BG_BASE,
+    'saybutton.text-color': themeColor.TEXT,
+    'oocbutton.background-color': themeColor.BG_BASE,
+    'oocbutton.text-color': themeColor.TEXT,
+    'mebutton.background-color': themeColor.BG_BASE,
+    'mebutton.text-color': themeColor.TEXT,
+    'asset_cache_browser.background-color': themeColor.BG_BASE,
+    'asset_cache_browser.text-color': themeColor.TEXT,
+    'tooltip.background-color': themeColor.BG_BASE,
+    'tooltip.text-color': themeColor.TEXT,
+    'input.background-color': themeColor.BG_SECOND,
+    'input.text-color': themeColor.TEXT,
+  });
 };
